SCUBA2: rename CSO entry as tau225
The CSO heading in the log files was causing some confusion, as it's
not actually derived from the CSO monitor values any more. Instead,
use a more generic heading called tau225 to be more explicit about
what it represents. Also update the corresponding POD to clarify that
the optical depths are those at the zenith.