SCUBA2: move all config adjustments in here
Rather than keep adding special exceptions in the makemap-calling
primitives, it makes more sense to put all the config-fiddling in one
places. Now this primitive returns a complete useable config string
consisting of a config file with override parameters specified as a
comma-separated list.

Experience shows that simply tacking on overrides is risky as it will
fail if the string is too long. However, most of the overrides are
exclusive so the risk should be low.

Note that overrides are only applied to system config files - if the
user provides a config file, it is used as is.