SCUBA2: retrieve and store NEP results for CHECK_RMS
As part of the CHECK_RMS analysis, the average NEP is required. This
primitive already calculates the noise/NEP and so can be easily
modified to deal with this requirement.

When carrying out CHECK_RMS analysis, the input data are tagged as raw
so they can be retrieved for the call to the map-maker. After the
noise calculation, the NEP statistics are retrieved from the NEP index
object and an average value calculated and stored in the Frame uhdr.

Note that the noise calculation for Mars is skipped as Mars is too
bright to make trustworthy noise estimates.