SCUBA2: new recipe to perform offline CHECK_RMS processing
This recipe is designed to be run offline (batch mode) on complete
observations to perform full timeseries and image-plane analysis of
the data to compare with the integration time calculator.

Results are written to log.checkrms in the same format as for the
PICARD recipe and the SUMMIT pipeline recipe.