PICARD: use hdrval as OBJECT might be in subheaders
The Frame object may now contain files with different OBJECT keywords
as it is updated immediately before checking to see if the images are
to be registered. The iscalsource() method seems to return true if the
argument is undef, as is the case when OBJECT is in the subheaders
(see issue #5). I wasn't expecting that.