SCUBA2: don't assume offsets are always in degrees
For the vast majority of cases it is true that the offsets will be in
degrees for this check - it's only if a non-pointing/focus recipe is
being used to process the pointing/focus data that the offsets are not
guaranteed to be in degrees. Do a quick check for sexagesimal format
and obtain values in degrees.