SCUBA2: add support for processing pointing observations
This primitive was originally written to only consider science
observations. Howevwe, now that it has been adopted in a different
recipe, it should support any on-source data - namely pointings.

Pointing observations are not contiguous in time so calcnoise produces
one output file per subscan. However, in order to calculate the
average NEP correctly, it is necessary to process each subscan
separately.