SCUBA2: define open-shutter but non-moving as not "on sky"
The absence of positional headers means that things like applying an
extinction correction to get a zenith value cannot be carried
out. Therefore, check for defined translated airmass and classify
those data from non-moving sequences as not "on sky" and let the
existing logic print an undefined value for the zenith NEFD in the log
file.