SCUBA2: select files with observational data if necessary
For some noise observations, the Frame can end up with an on-sky noise
sequence where the telescope is tracking and another where it is
not. That fact that these are different sequences (even if they are of
the same kind of data) means that sc2clean can't concatenate the two
files to make a single noise calculation. Therefore, given a choice,
the pipeline now selects only the data associated with the observation
(where the SEQ_TYPE and OBS_TYPE headers are the same).