Calib: protect against empty matches
This fix comes from a PICARD recipe that mosaicked different sources,
which led to an undefined value being passed to the JCMT isplanet()
method. Upper-casing an undef value produces an empty string which
then matches true against the list of planets, leading to all manner
of undesireable consequences.

Fixes issue #5.