heterodyne: Ensure that the output Frame has the retained suffix.
The _tsmsk suffix is retained by recipes for the iterative group
stage.  A Frame with this suffix does not get created if there is no
mask associated with the survey, but for DAS observations the 'survey'
is the instrument name and at present has no associated mask.  Even if
this omission may change, it is prudent to make this modification to cope
whenever no masking is required.

This change copies the input to _tsmsk even if there is no masking to
perform.  The NDFCOPY was already incorrectly listed in the used tasks.