SCUBA2: only display results every third subscan in QL
This is a simple attempt to reduce the amount by which the QL pipeline
slows down over the course of a night's observing. The slowdown is
caused by the increase in the size of the AGI database, and calling
the display less often results in a smaller file, and thus shorter
display times.