Core/Print/Xorac: add ORAC "results" box to X GUI window
This commit introduces the concept of an ORAC "result", which is a
value calculated by the pipeline that is meant to remain visible to
the user for longer than the typical duration of processing an
observation rather than get lost in all the other informational
messages written by the pipeline.

The new command is "orac_result", results are printed in green text,
and only printed when the X GUI is used (i.e. -log x). Calls to
orac_result are ignored if writing the output to a terminal window
(-log s) since there is currently no way to control text output to
terminals.

orac_result is meant to be used sparingly, and output should be as
concise as possible to maximize the use of the available space.

The box for results is the same size as the current warning and error
boxes, but could be enlarged if desired.