heterodyne: Handle the special case of abnormal tuning for a standard.
Sometimes the standard is not retuned correctly.  The nominal
velocity range of the emission line to be measured is consequently
not within the velocity range of the spectrum.  This fix prevents
the COLLAPSE flux integration from dying because of this.  It also
does not compute the peak and no record written to log.standard.
Instead a warning message appears.