SCUBA2: issue errors if noise and/or flatfield results are absent
With these changes, a SETUP will only be deemed successful if the
fastflat *and* noise calculations succeeded (previously the check on
valid fastflats was not presetn). If either of these criteria is not
met the pipeline will print an error and write a dummy flag file,
forcing the SETUP to be repeated.