SCUBA2: remove unused refimage method and attribute
This method and associated attribute was added a long time ago with
the intention of implementing something useful. In practice, the
calibration object isn't the best place for storing a reference image
as there is the risk that it is picked up and used by different
Groups, a scenario that is unlikely to end well.