SCUBA2: add refimage method for storing/retrieving reference images
A reference image is useful for defining coordinate systems when
mosaicking or coadding images. This method stores the name of files
that may be used as references and allows separate files to be stored
for celestial or focal-plane mosaics as both types are needed in the
pipeline. The Group uhdr is used rather than add extra object attributes.