SCUBA2: always create reference for focal-plane mosaics
The QL pipeline needs all the help it can get to keep processing time
to a minimum, and creating a reference image each subscan
unnecessarily adds a second or two. Not only that, it turns out that
the noise/NEP mosaics weren't even using a valid reference, instead
always using the first file in the Frame. The whole point of having a
reference is to ensure that the mosaics are always oriented the
correct way, which is not guaranteed when using one of the subarray
noise images as a reference.

Therefore always create a reference (if one does not already exist) if
the mosaic is in focal-plane coordinates even for science sequences so
the QL will be able to reuse a suitable reference.