SCUBA2: store the current mosaic as the reference if required
When creating the focal-plane mosaics in the QL and SUMMIT pipelines,
the reference image needs to persist over multiple sequences. Rather
than leave a series of dedicated reference images on disk, it is
preferable to use a persistent output file such as the noise/NEP
mosaic. Therefore, store the current mosaic as the reference - unless
a reference has already been stored. This is required to avoid the NEP
mosaics being stored as the references as they are deleted at the end
of the recipe.