SCUBA2: protect against storing missing files in the Frame
This primitive assumes that it has data for each SMU position, and in
particular that the map for the central (null) SMU position is
present. There have been cases in the QL where data were skipped or
took long to process such that the central map was not present. Things
quickly go wrong when a non-existent file is added to the Frame and
its header read.

So now check that the centre file exists, and if not further check for
one of the neighbouring positions. If one of those exists then it is
used to determine the source peak position for shifting the WCS of the
cube. If none exist then the WCS is left unadjusted. The end result is
that the old situation should never arise and the QL has one less
reason to fall over.