SCUBA2/general: generalize and rename _ADD_IMAGES_ to _ADD_IMAGE_COMPS_
To work with PICARD, _ADD_IMAGES_ has been generalized and renamed to
_ADD_IMAGE_COMPS_. The new name is a more accurate reflection of what
the primitive does: it adds together named NDF components and writes
the sum to the corresponding extension in a given output file.

All traces of SCUBA-2-specific code have been removed - the only
difference is that the primitive now must be told the extension name
and corresponding NDF components.

Despite looking like a delete-and-add-new commit, it's really a rename
and edit so I've left the contribution history intact.