SCUBA2: only dereference when necessary
No point checking things twice - check the return value from the
primitive and dereference within that code block if the hash reference
exists, rather than dereferencing it if possible and then checking if
the hash is defined.

Also expand the comment for the case when no fastflat solutions were
retrieved.