ifu: Look for any _im to be image product.
This is to overcome the reuse of this primitive where the suffix
can be set to _im_fc, i.e. flux corrected, but in fact the flux
correction has not occurred.

The same applies to _COLLAPSE_CUBE_TO_FORM_SPECTRUM_ but for the
spectrum product in Commit 1a022ee5554b8734c, but I missed it
amongst other changes while half awake at LAX.