SCUBA2: calculate NEFD and trim image before analysis
The NEFD is calculated in a later primitive, but if left until then
the trimming primitive complains about a missing NEFD component. In
the interests of not confusing and worrying users unecessarily, just
calculate the NEFD befoer trimming the image.

Oh yeah and trim the image to a 90-arcsec radius circle before making
measurements! Otherwise PICARD and ORAC-DR will never agree...