general: Align vector plot with corresponding map
For the vector plot to be aligned with the corresponding image,
it needs to be produced with the same "device".  Therefore changed
the plot device from transparent PNG to the plain device type
and removed the restriction that it be PNG.  Aligning the
images also required them to be of the same size, and to have
the key turned off for the vector plot.  (Otherwise polplot
fails saying that there is no room for the key.)  We don't
have keys for normal thumbnail images and spectral, so it
makes sense to also not have the key for the polarimetry
vectors.