imaging: Make greyscale previews if in multicolor
The "NCOLOUR" recipes produce multiple preview graphic files with
the same product ID.  These need to be merged by ukirtwrapdr
because there can only be one preview per plane in CAOM-2 and
all files with the same product ID tend to go in the same plane.
Therefore in multicolor modes the pictures need to be greyscale
so that ukirtwrapdr can color-composite them easily.