lib: Make oracdr-locations file non-group-writable
This is to prevent the pipeline from undoing the action
described in the third response to fault 20140110.004.
When I logged in at night to investigate fault 20140201.002,
I was alarmed to find that the name of the temporary
directory in which I had been testing the pipeline had been
scribbled over the corresponding file in oracdr-locations,
which must have happened because ORAC-DR running as operator
had made it group writable again.