SCUBA2: use the base position for the map centre
The WCS skyref attribute can be different from the base position
depending on whether an external reference or mask was specified at
the time the map was made, and is therefore not reliable as an
indicator of the map centre position. Instead, use the base position
and take account of any offsets, which is the same as the
AST-region-based masking.

The situation for Group files isn't as simple as the base position
might be different for each file in the coadd. For now, then, this
primitive only operates on Frame objects. Of course when called as
part of a PICARD recipe there's no guarantee the base position will be
correct, but that's an issue for another day.