spectroscopy: Remodel the creation of the tagged product.
The aim is to obtain a group file of the form g<i><utdate>_<group>_arc
for the sciene-archive wrapdr script rather than Frm equivalent.  Use
a temp Grp file and let _TAG_AS_REPRESENTIVE do the most of the work,
passing it the arc as the input file and then request an update of the
Grp file.  The previews are derived from the suitably named product.