SCUBA2: store observation number and flatfield solution in FLATS
The observation number was being stored for the case when the
fastflats are processed on their own, but not the case when the
fastflats were being processed with the current data (e.g. the SUMMIT
pipeline) where the relevant information is retrieved from the hash
reference stored in the FLATS uhdr entry.

In addition, SETUP flag files contained the wrong flatfield file, and
now store the flatfield solution rather than the raw file.

Both of these fix batch processing of SETUPs.