spectroscopy: Redesign 3 of products and their names.
After further discussion with Graham and Watson we have decided the
following changes.

1) The wrapdr script will expect preview graphics to have names of the
form g<i><utdate>_<grpnumber>_<product>_[rsp|rimg]_<size>.png.
Products continue to have the form g<i><utdate>_<grpnumber>_<product>.

2) There is one special case.  For the reduced product the file names
have the "fc" suffix for <product>, but the PRODUCT keyword in the FITS
and the jsa:product keywords in the preview files remain "reduced".

Under the previous scheme it was possible to end up with no (reduced)
previews if the flux calibration fails.  By putting each of the
products in a separate plane each can have a preview.  Some will be
rsp and some will be rimg, but it seems not both.  The rimg is for
extended-source or long-slit data.

These modifications are for ordinary spectroscopy.