ifu: Redesign 3 of products and names.
See d51e40d16bbda for the name format changes.  We have four
products, each with preview images.

  product          suffix           purpose
  ------------------------------------------------
  stackedspectra   no ext / _dif
  spectralcube     _cube            first reformatted cube with
                                    spatial and a spectral axis
  dbs              _dbs             divide by standard
  reduced          _fc              flux calibrated

This is like spectroscopy which has an aligned product with an _aws
suffix instead of the spectral cube and the no ext product is wavecal.
[Noted here as I forgot to tabulate like the above for d51e40d16bbda.]

The last three can have _im white-light and _sp mean spectrum
representations in the same plane.

There is a known occasional problem in display mode where an
AGI__DIFDV error appears claiming DISPLAY is changing device from
PNG to GWM while it is attepmting to make a preview image.  Since this
does not prevent testing at CADC it can be investigated further later.