ifu: Shortened product name to prevent DISPLAY anomalies.
At the moment I don't understand this.  Story is without this change
the pipeline falls over trying to make the previews.

  Err: !! The default position was accepted for parameter CENTRE ('1.9535
  Err: !     20:24:25.5') but cannot be used!
  Err: !! Aborted attempt to obtain a position using parameter CENTRE.
  Err: !  DISPLAY: Failed to display an image of a one- or two-dimensional data set.
  Err: !! PAR__ABORT: Parameter request aborted
  Err: Error in obeyw to monolith kapview_mon (task=display): 146703171

Graham has also seen an error saying the combination of CENTRE, XMAGN,
and YMAGN accesses no good data.  I had also seen an AGI__DIFDV error
claiming a switch from PNG to PS when trying to create the PNG for
spectroscopy, not ifu as I had initially thought.  (It was a mixed
night.)

Since the _TAG_AS_REPRESENTATIVE_ line looked just like any of the
others changed in the same commit but didn't cause the pipeline to
crash I looked elsewhere.  The only other significant change to this
primitive in 77cf6697a8d4947 was the product name.  Shortening fixes
the immediate problem, but it suggests that the long preview name is
somehow messing with the DISPLAY parameters, which is worrisome.