heterodyne: Create obs products for continuum recipe.
If an observation is marked as not good it is processed in obs mode.
The recipe as was failed to created obs products and previews, thus
dpCapture finds no files to ingest into the JSA.  I imagine that obs
products were not made because the signal-to-noise would be too low
and the recipe sums these in the group file.  So this change trades
generation of unwanted noisy obs products, versus getting no products
at all if the observation was questionable or bad.

It would be better to create the obs products only when necessary,
perhaps when there is a single observation in the group.