heterodyne: Work around WCSALIGN insitu problem.
This addresses Fault 20130621.006.  Until the source of the
non-integer shifts between subbands is identified---original data or
GSD2ACSIS---this provides a workaround which assumes that the shifts
are small.  In effect it finds the nearest-neighbour pixel offsets and
sets the origin of the second file.  The result is that the composite
spectrum created by merging does not show the joins between the
subbands.  Before this fix the wrong regions were compared, leading to
incorrect signal offsets between the bands and jumps in the spectrum
in the overlapping regions.