SCUBA2: check subheaders exist before accessing attributes
The SUMMIT pipeline was crashing in some instances where it was
processing a mixture of subscans. If the Frame contained data from
different sequence types, and the science data had already been
accounted for in terms of the observation headers, then the
$startsubhdr and $lastsubhdr subheaders were undefined. This caused
DateTime to throw an error when trying to parse an undefined string,
thus causing the pipeline to crash fatally.

The solution is to check that those subheaders are defined before
trying to access them.

Note that it doesn't matter if a map is produced from these science
data and an additional entry made in log.checkrms: all the values
derived from the header will already be OK, and the map-based
parameters will now be from a map with all the data included.