SCUBA2: check for non-zero standard deviation
Sometimes makemap produces a blank map full of zeroes. This causes
problems downstream when attempting to find sources - findclumps
complains about zero variance. Therefore, since this primitive is
already calculating the statistics to check for good bolometers, it's
trivial to add another check for a non-zero standard deviation.

In the case of multiple output tiles, each tile is checked and in turn
and only those that pass are analyzed.

If the output map(s) (or tiles) has (have) zero standard deviation then
the Frame object is marked as bad.