JSA: add separate tiles preview primitive
This primitive creates a preview graphic for each tile, which
has been requested for testing the JSA tile making recipe at
JAC.  These previews are only made if the TILE_PREVIEWS
recipe parameter is set to true.

The tiles are now made at full size (not trimmed) so that corresponding
tile previews are the same size. Since the products are already being
trimmed when they are tagged, they should not be changed.