SCUBA2: avoid specifying makemap config twice
Since _GET_MAKEMAP_CONFIG_ reads the MAKEMAP_CONFIG recipe
parameter, _GET_MAKEMAP_ARGS_ doesn't need to include it
in the set of arguments it returns.  The other way to
solve this would be to remove "config" from the list of
arguments which ORAC::Frame::SCUBA2::makemap_args()
returns but then _REDUCE_SCAN_STEER_ would have to
handle it specially because it also uses the
RECPAR_MAKEMAP_CONFIG uhdr entry.