PICARD: add a recipe to co-add JSA tiles
This commit adds a new recipe COADD_JSA_TILES and corresponding
primitive _COADD_JSA_TILES_.  These form sub-groups using the
JSATILE header.  Then if there are multiple files for a tile, they
are co-added using MAKEMOS with METHOD=MEAN and USEVAR turned on.

A new recipe name was used so that there is a specific recipe
which can be used for generating "pro" and "pub" co-adds from the
corresponding observation tiles in the JSA.  It is possible that
there will also be additional processing at this stage, so it can
be added to this recipe if needed.  (Smoothing and calibration
have been discussed as steps which could possibly be done here.)
It will also be more convenient for offline testing if there is a
specific recipe for this.

The output files should be suitably-named for jsawrapdr to prepare
them for archiving.  Since we do not yet have an agreed naming
scheme, a placeholder including a dummy ASN_ID is currently used.