JSA: use new inout_jsatile method to tag tiles
When _TAG_AS_REDUCED_PRODUCT_ determines the new filename to use,
it now checks for the JSA_TILES uhdr entry.  If this is set then
a new inout_jsatile method (of ORAC::JSAFile) is called instead
of the normal inout method.  This new method appends the JSA
tile number (from the JSATILE header), zero-padded to six digits,
to the suffix to generate a JSA-style file name.  It also makes
use of the known filename pattern to know how many filename
components to keep.

To ensure that this works reliably, the SCUBA-2 primitive
_MAKEMAP_FRAME_ ensures that the JSA_TILES header is set when it
runs in JSA tiles mode, and _FUDGE_CALIBRATOR_POSITION_ ensures
that its preliminary call of _MAKEMAP_FRAME_ does not run in
JSA tiles mode.