PICARD: use inout_jsatile to tag co-added tiles
Using this new method prevents the accumulation of suffices in
the co-added tiles generated by COADD_JSA_TILES.  To do this it
is necessary to load the instrument frame classes because the
PICARD frame class can't inherit from JSAFile.  Having the frame
classes will also help in determining the association information
used in the filename.  For SCUBA-2 this is now just the wavelength.