lib: Return plain filenames from TempFile
This is to try to reduce the chances of encountering the message
length limit.  Apparently full temporary file pathnames have been
a problem in the ACSIS pipeline for a while, and they have
now become a problem in the SCUBA-2 pipeline since the introduction
of the OUTFILES parameter to the MAKEMAP invocation.