SCUBA2: include SIGMA in list of initialized values
Log file entries with no good bolometers had one fewer element on the
line than those with good bolometers: the value for SIGMA was not
defined, which caused problems with parsing the log files. The problem
was caused by the lack of an entry for SIGMA in the initial QA results
info hash. Add it to the list of zero-initialized values and the issue
is solved.