heterodyne: Correct the rotation angle crota.
It was generating a wrong sign for crota for DAS observations (Fault
20140315.002).  Not sure why the 90-PA was there as it with the
subsequent loop constraining angles between -45 and 45 it effectively
flips the angles.  Note that it has not been tested on genuine ACSIS
data, where perhaps the flip is necessary.  If this is the case, we
can put a BACKEND test, or make GSD2ACSIS behave like ACSIS.