SCUBA2: account for cases near the northern meridian
When tracking a source across the northern meridian, the demand and
actual azimuth values can be either side of the line leading to
apparent differences of many degrees and average values in the region
of 180. The pipeline reports errors when it sees such values. The
solution is to check for the situation when the telescope is tracking
very close to the meridian and perform modulo 360 arithmetic as
necessary.