SCUBA2: include "um" units in subsystem_id
Russell has required that the productID for SCUBA-2 include
the filter with wavelength, i.e. it is to say 450um or 850um.

I separated the implementation of file_from_bits_extra()
from subsystem_id() to allow them to be edited separately.
The alternative would have been to redefine subsystem_id()
in terms of file_from_bits_extra(), but that would cause
the productID to inadvertently change if file_from_bits_extra()
is updated.