ACSIS: add freq and bwmode to subsystem_id
Russell has required that the productID for ACSIS include the
rest frequency and bandwidth mode.

As for the previous commit, for SCUBA-2, the implementation of
file_from_bits_extra() has been separated back out from
subsystem_id() to allow them to be edited independently.
The rest frequency determining code from findgroup() has
been moved to a new method so that it can be re-used
by subsystem_id() but this required an additional option
to prevent rounding.  (The previous behavior for group
association IDs should be preserved.)