SCUBA2: also use matched-filtered map to assess whether a peak is detected
For a genuine source detected at low signal-to-noise, applying the
matched filter should more or less double that signal-to-noise
ratio. Therefore a way to test if an image is blank is to check to see
that the peak S/N in the matched filtered map is indeed higher than in
the signal map. Blank maps have a largely unchanged peak S/N.