SCUBA2: extract _CHECK_INBEAM_HEADER_ primitive
Extracted the FTS-2 and POL-2 checks of the INBEAM header
from the _FTS2_HELLO_ and _POL2_HELLO_ primitives and placed
them in a new _CHECK_INBEAM_HEADER_ primitive.  This primitive
is then used in the REDUCE_SCAN_JSA_PUBLIC recipe to ensure
that observations where the INBEAM header says that either
FTS-2 or POL-2 were in the beam are not reduced to form the
JSA public HEALPix tiles.  This change was requested by
SFG.