JSA: group also by ASNID in _COADD_JSA_TILES_
This primitive now goes through all the frames during the frame
stage and adds the JSA public association ID as a uhdr.  This
uhdr has been added to the subgrps keys list to ensure that
tiles from different associations are not co-added.  To do this
the code for selecting the frame class has been extracted as
a subroutine, and since the group stage no longer needs to
determine the backend, the instrument-specific co-add recipe
is now chosen by the primitive search path mechanism rather
than being selected explicitly based on the backend.