JSA: create untrimmed tile previews for co-adds
When generating preview images for co-added JSA tiles, the tiles are
first untrimmed to create a temporary file which covers the whole
area of the corresponding tile.  This temporary file is then used to
generate the preview image only.  The reasoning behind this is that
the JSA tile is notionally the whole predefined HEALPix tile, and
therefore this is what we should be showing a preview of, regardless
of the fact that we may store trimmed data arrays for efficiency.
The useful result should be that all the tile previews for each
instrument should be the same scale.