JSA: tag JSA tile co-adds in group context
Co-added JSA tiles are now tagged as the reduced product (actually
"healpix") in group context so that the pipeline sets the ASN_TYPE
header to "night" (for jsawrapdr to correct later) and adds the
ASN_ID header.  This required the creation of a temporary group
and improved intermediate tracking.  The tagged product is now
added to the list of intermediates so the keep pattern needed to
be corrected from "reduced" to "healpix" to prevent the products
from being deleted.