SCUBA2: rewrite to use more robust method using matched filter
This primitive now uses the method first included in _FIND_MAP_PEAK_
where the matched filter is applied to the image to search for a
peak, which will show up as a more significant maximum if present.

The criteria for a peak are now that the peak S/N in the signal map
must exceed 3, the peak S/N in the matched-filtered map must exceed 5
and the matched-filtered peak S/N must be greater than the signal S/N.

If all three are satisfied, the position of the peak S/N is returned
to the caller in the primitive hash.

Note, however, that if given a map that has already been processed
with the matched filter, the only criteria used is that the peak must
exceed 5-sigma. This assumes that if true, the original map must have
had a minimum 3-sigma peak.

If a fit to the source fails, the value at the peak position and its
associated uncertainty are returned.