SCUBA2: use only flatfielded data in the SUMMIT pipeline
In the SUMMIT pipeline a map should only be made using files with
flatfielded data, if flatfielded data exist. In this case fastflat and
noise data should not be included in the list given to makemap. (Note
that if no flatfielded files exist, then the map is being made from
all of the raw data taken thus far and it's OK to include the
fastflats.)