SCUBA2: call peak-finder directly on matched-filtered map
At some point in the past I thought that calling _FIND_FLUXES_ was the
right thing to do for both the signal and matched-filtered maps. Now
it seems unnecessary for two reasons: 1) the running code to derive
results that aren't used and 2) from a logic perspective: it's now
obvious that it is only the peak of the matched-filtered map that is
useful.