ACSIS: alter frame rest_frequency method behavior
The rest_frequency method previously preserved the historical
behavior of the code which was extracted from the findgroup method.
This is now an option, but not the default.  By default it now only
looks for the actual rest frequency in the header or in the WCS.
Averaging the FRQSIGLO and FRQSIGHI headers is only permitted if the
argument ("$allow_sig_av") is set to a true value.  The previous
argument (to prevent rounding when these frequences are averaged)
is no longer required, and so has been replaced by the new one.
The argument used in all calls to rest_frequency has been swapped
to take account of the new default.