lib: add support for PRODID header
Russell has suggested this header be added, corresponding to the
CAOM-2 plane productID field.  The pipeline already has a product_id
method which must be used to give preview images the same product
ID as will be used in CAOM-2 in order for jsawrapdr to be able to
rename them appropriately for the the CADC web interface to be able
to find them.  This method can therefore be used to set PRODID
at the same time as the PRODUCT header is set.

Edited code which referred to PRODUCT as the "product ID" in order
to avoid confusion with PRODID.  The default comment is now "Pipeline
product", as used by ORAC::BaseNDF::collate_headers.