PICARD: replace RA,DEC parameters with generic X,Y for registering images
To allow images to be registered in Galactic coordinates, the
REGISTER_RA and REGISTER_DEC recipe parameters have been replaced by
corresponding X and Y parameters.

A naive test is used on the X,Y values to determine the coordinate
system: if the X and Y values contain a colon, then it is assumed the
coordinates are J2000, otherwise Galactic.

Note that now the X,Y position is stored in the Frame uhdr as an array
reference, not as separate entries.