SCUBA2: avoid using _MAKEMAP_FRAME_'s old $out
$out is the requested MAKEMAP output filename but @out is the list
of files it actually produced.  Outside of JSA tiles mode, @out
would normally be ($out) and so the block which tests the SkyRefIs
attribute of $out for pointing and focus observations normally works
because these observations aren't normally reduced onto JSA tiles.
This commit sets $out to $out[0] inside this block to be sure that
it is checking a file which MAKEMAP actually created.