heterodyne: Select the correct files for group processing.
This addresses Fault 20140506.001 where the creation of the group cube
fails because it is delivered a reduced cube rather than a time-series
cube via Frm->file of Grp->members.  Use a tag to restore the time
series as the current file at the end of the recipe.  This at least
sorts the mechanics, but there may be a better fix in one of the
primitives, as it's unclear what change caused this error. It's
not a commonly used recipes so it might have been disfunctional for a
while.