JSA: avoid trying to write PRODID header if empty
The code in _TAG_AS_REPRESENTATIVE_ uses an empty string for the
product ID if it can not be determined.  However trying to set
a FITS header to an empty string with _SET_FILE_FITS_ITEM_
causes a problem.  Therefore this commit guards the call to
that primitive (and the vector catalog equivalent) so that
it is avoided when the header would be empty.