SCUBA2: allow different reference and image coordinate systems
The input coordinates do not have to be specified in the same
coordinate system used in the images to register.

The coordinate system of the shifts applied is now reported with the
offsets just to be clear which system is being used.

However, no check is made (other than the initial attempt to fit the
peak) to ensure the given position lies within the bounds of the
image. (That has always been the case - it just seems more likely to
cause problems when mixing coordinate systems.)