PICARD: add primitive to merge FITS headers
The new _MERGE_FITS_HEADERS_ primitive uses the SMURF task FITSMERGE
to merge the FITS headers from various files and write the result
into another file.  This primitive is now used in the SCUBA-2
_COADD_SINGLE_JSA_TILE_ primitive so that the co-added tile has
merged FITS headers.  The call has also been added to the commented
out bit of the ACSIS _COADD_SINGLE_JSA_TILE_ primitive so that it
is ready for when that primitive is implemented.