heterodyne: Enable user-defined trimming limits.
There are arguments and parameters to set the limits, but these were
mostly ignored and the advanced method used.  Now if a user sets the
limits via recipe parameters or invokes the primitive with the LOWER,
UPPER, or PERCENT arguments, that's what will be clipped.

Also set the defaults by backend when only one of the lower and upper
bounds has been specified.

Added some documentation to clarify when the parameters and arguments
take effect.