general: move from SCUBA-2 to general
This primitive contained no SCUBA-2-specific code and could therefore
be used by any instrument for determining the statistics of an
image. Moving it into the general directory makes it available for all
instruments, including PICARD.

Despite the name, the primitive can be used on multi-dimensional data,
not just images.