general: use a generic name for calculating SNR images
Since this primitive contained virtually no SCUBA-2-specific code, it
can be easily generalized for wider use.

The only SCUBA-2-specific element was knowledge of the WHITEN_BOX
Group uhdr entry to override the region over which statistics are
calculated. This has been removed and the relevant caller updated to
specify the region explicitly.