SCUBA2: use ARCSEC FCF for co-added JSA tiles
The SCUBA-2 _COADD_SINGLE_JSA_TILE_ primitive now sets the argument
UNITS=ARCSEC when it calls the _CALIBRATE_DATA_ primitive.  This should
result in the co-added map not having per-beam units.

Note: it still uses the default FCF for the HEALPix tiles, pending
confirmation or determination of a new FCF once the generic
map making parameters are determined and the calibrators are
analyzed following reduction with them.

Note also: the determination of the pixel size could be simplified
for HEALPix tiles since we should know how big they are.