lib: do not include PICARD instrument in log name
When ORAC_LOGDIR is set, an extra HTML pipeline log file is
written to a file in that directory, the name of which includes
the environment variable ORAC_INSTRUMENT.  However PICARD sets this
variable to the string "<not required>", leading to unsuitable log
file names.  This commit updates the block which determines the
name of this log file so that it recognizes this string and leaves
it out of the file name.