PICARD: loosen domain name check for JAC domain
sc2dr5 suffers the same problem reported in fault 20140327.003
with an extra dot being present at the end of the domain name.
Then when you ran SCUBA2_CHECK_RMS, _SCUBA2_GET_NEP_FROM_FILE_
couldn't find the NEP log file because it didn't know this was a
JAC machine.  This commit removes the $ from the end of the match
to work around this.