general: set pixels with anomalously-low variances to bad
Pixels with anomalously-low variances are given disproportionate
weight when mosaicking images and then show up as oddities in the
final map. This primitive is based on an algorithm outlined by David
Berry. First it forms a "weight" image by taking the reciprocal of the
variance, before calculating the statistics with successive clipping
at 5, 5 and 5 sigma. The maximum value in these statistics is used as
a threshold, converted back into a variance (by taking the reciprocal)
and the variance component of the original image is thresholded at
that value.

Creates a new image with suffix "_maskv".